+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/label/package-info.java
@@ -17,10 +17,7 @@
  * Copyright (c) 2023 - 2025, Björn Darri Sigurðsson.
  */
 /**
- * Package configuration values:
- * <ul>
- * <li>{@link is.codion.swing.common.ui.component.label.LabelBuilder#HORIZONTAL_ALIGNMENT}
- * </ul>
+ * JLabel related classes.
  */
 @org.jspecify.annotations.NullMarked
 package is.codion.swing.common.ui.component.label;